+++ app/src/commonMain/kotlin/model/Point.kt
@@ -20,6 +20,6 @@
 package model
 
 data class Point(
-    val x: Int,
-    val y: Int
+    val x: Short,
+    val y: Short
 )
